# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: HITL Protocol API
  version: "0.5"
  description: |
    The Open Standard for Human Decisions in Agent Workflows.

    HITL Protocol standardizes three-party flows between **Agent**, **Service**, and **Human**.
    When an agent calls a service API and the service needs a human decision, the service returns
    HTTP 202 with a `hitl` object. The agent forwards the review URL to the human. The human
    decides in a browser. The agent polls for the result.

    **Framework-agnostic.** This spec works with any HTTP server (Express, Hono, Next.js, FastAPI,
    Spring Boot, Go net/http, or any other).

    For the full specification, see: https://github.com/rotorstar/hitl-protocol
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: HITL Protocol
    url: https://github.com/rotorstar/hitl-protocol
servers:
  - url: https://api.example.com/v1
    description: Example service API
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Agent-to-service authentication. The agent authenticates with a bearer token
        obtained through the service's standard auth flow.
    ReviewToken:
      type: apiKey
      in: query
      name: token
      description: |
        Opaque bearer token in the review URL (43 chars, base64url-encoded).
        Generated by the service: `randomBytes(32).toString('base64url')`.
        Verified by the service: `timingSafeEqual(SHA256(token), stored_hash)`.

  # ============================================================
  # Headers
  # ============================================================
  headers:
    Retry-After:
      description: Seconds until the agent should poll again. RECOMMENDED 30-300.
      schema:
        type: integer
        minimum: 1
        examples: [30]
    ETag:
      description: Opaque identifier for the current poll response version. Enables efficient polling with If-None-Match.
      schema:
        type: string
        examples: ['"v3-completed"']
    X-RateLimit-Limit:
      description: Maximum requests per minute.
      schema:
        type: integer
        examples: [60]
    X-RateLimit-Remaining:
      description: Remaining requests in the current window.
      schema:
        type: integer
        examples: [58]

  # ============================================================
  # Parameters
  # ============================================================
  parameters:
    caseId:
      name: caseId
      in: path
      required: true
      description: Unique, URL-safe review case identifier.
      schema:
        type: string
        pattern: '^[a-zA-Z0-9_-]+$'
        examples: ['review_abc123']
...
    ifNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: ETag from a previous poll response. Returns 304 if unchanged.
      schema:
        type: string
        examples: ['"v2-in_progress"']

  # ============================================================
  # Schemas (referencing existing JSON Schema files)
  # ============================================================
  schemas:
...
    PollResponse:
      $ref: './poll-response.schema.json'
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Machine-readable error code.
          examples: ['duplicate_submission', 'rate_limited', 'invalid_token', 'case_expired']
        message:
          type: string
          description: Human-readable error message.

  # ============================================================
  # Response Objects
  # ============================================================
  responses:
    NotModified:
      description: Poll response unchanged since last ETag. Agent should wait and retry.
...
    RateLimited:
      description: |
        Too many requests. RECOMMENDED limit: 60 requests per minute per case.
      headers:
        Retry-After:
          $ref: '#/components/headers/Retry-After'
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: rate_limited
            message: Too many requests. Please wait 30 seconds.
...
paths:
...
  /reviews/{caseId}/status:
    get:
      tags: [Polling]
      summary: Poll review case status
      description: |
        Agent polls this endpoint to check the current status of a review case.
        Returns one of 6 statuses: `pending`, `opened`, `in_progress`, `completed`, `expired`, `cancelled`.

        **Efficient polling:** Use `If-None-Match` with the `ETag` from a previous response.
        If the status hasn't changed, the service returns 304 Not Modified with no body.

        **Polling interval:** RECOMMENDED 30 seconds to 5 minutes. Respect `Retry-After` header.
      operationId: pollReviewStatus
      parameters:
        - $ref: '#/components/parameters/caseId'
        - $ref: '#/components/parameters/ifNoneMatch'
      responses:
        '200':
          description: Current review case status.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollResponse'
              examples:
                pending:
                  summary: Case created, human hasn't opened yet
                  value:
                    status: pending
                    case_id: review_abc123
                    created_at: "2026-02-22T10:00:00Z"
                    expires_at: "2026-02-23T10:00:00Z"
                opened:
                  summary: Human opened the review URL
                  value:
                    status: opened
                    case_id: review_abc123
                    created_at: "2026-02-22T10:00:00Z"
                    opened_at: "2026-02-22T11:15:00Z"
                    expires_at: "2026-02-23T10:00:00Z"
                in_progress:
                  summary: Human is interacting (multi-step form progress)
                  value:
                    status: in_progress
                    case_id: review_abc123
                    created_at: "2026-02-22T10:00:00Z"
                    opened_at: "2026-02-22T11:15:00Z"
                    expires_at: "2026-02-23T10:00:00Z"
                    progress:
                      current_step: 2
                      total_steps: 3
                      completed_fields: 4
                      total_fields: 9
                completed:
                  summary: Human submitted their response
                  value:
                    status: completed
                    case_id: review_abc123
                    created_at: "2026-02-22T10:00:00Z"
                    opened_at: "2026-02-22T11:15:00Z"
                    completed_at: "2026-02-22T11:19:46Z"
                    expires_at: "2026-02-23T10:00:00Z"
                    result:
                      action: select
                      data:
                        selected: [job_001, job_003]
                    responded_by:
                      name: Alex Mueller
                      email: alex@example.com
                expired:
                  summary: Case expired without response
                  value:
                    status: expired
                    case_id: review_abc123
                    created_at: "2026-02-22T10:00:00Z"
                    expires_at: "2026-02-23T10:00:00Z"
                    expired_at: "2026-02-23T10:00:00Z"
                    default_action: skip
                cancelled:
                  summary: Human cancelled the review
                  value:
                    status: cancelled
                    case_id: review_abc123
                    created_at: "2026-02-22T10:00:00Z"
                    opened_at: "2026-02-22T11:15:00Z"
                    cancelled_at: "2026-02-22T11:16:30Z"
                    reason: "Changed my mind about the application."
        '304':
          $ref: '#/components/responses/NotModified'
        '429':
          $ref: '#/components/responses/RateLimited'
...